Autoselect

+++ src/main/java/com/bernardomg/tabletop/palette/palette/model/persistence/Paint.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 the original author or authors
+ * Copyright 2019 the original author or authors
  * <p>
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -54,7 +54,7 @@ public class Paint implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false, unique = true)
-    private Integer           id               = -1;
+    private Long              id               = -1l;
 
     /**
      * Name of the entity.
@@ -66,7 +66,7 @@ public class Paint implements Serializable {
     private String            name             = "";
 
     @Column(name = "palette_id", nullable = false)
-    private Integer           paletteId;
+    private Long              paletteId;
 
     /**
      * Constructs an example entity.
@@ -101,7 +101,7 @@ public class Paint implements Serializable {
      *
      * @return the entity's identifier
      */
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -109,7 +109,7 @@ public class Paint implements Serializable {
         return name;
     }
 
-    public Integer getPaletteId() {
+    public Long getPaletteId() {
         return paletteId;
     }
 
@@ -118,7 +118,7 @@ public class Paint implements Serializable {
         return Objects.hash(id);
     }
 
-    public void setId(final Integer value) {
+    public void setId(final Long value) {
         id = checkNotNull(value, "Received a null pointer as identifier");
     }
 
@@ -126,7 +126,7 @@ public class Paint implements Serializable {
         name = checkNotNull(value, "Received a null pointer as name");
     }
 
-    public void setPaletteId(final Integer id) {
+    public void setPaletteId(final Long id) {
         paletteId = id;
     }
