+++ src/main/java/com/bernardomg/tabletop/palette/palette/repository/PaletteGroupRepository.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018 the original author or authors
+ * Copyright 2019 the original author or authors
  * <p>
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -26,6 +26,6 @@ import com.bernardomg.tabletop.palette.palette.model.persistence.PaletteGroup;
  * @author Bernardo Mart&iacute;nez Garrido
  */
 public interface PaletteGroupRepository
-        extends JpaRepository<PaletteGroup, Integer> {
+        extends JpaRepository<PaletteGroup, Long> {
 
 }
